+++ jsql/jsql_test.go
@@ -36,11 +36,12 @@ func TestSql(t *testing.T) {
 	if err := Init(); err == nil {
 		t.Error(fmt.Sprint(testErr, " Init must be return error"))
 	}
-	SetFileName("../files/test-jconf-error.json")
+	SetRoot("../files/")
+	SetFileName("test-jconf-error.json")
 	if err := Init(); err == nil {
 		t.Error(fmt.Sprint(testErr, " Init must be return error"))
 	}
-	SetFileName("../files/test-jconf.json")
+	SetFileName("test-jconf.json")
 	if err := Init(); err != nil {
 		t.Error(err)
 	}